% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contest_data.R
\name{get_contest_data}
\alias{get_contest_data}
\title{Obtain Shiny Contest Submissions}
\usage{
get_contest_data(year = 2021, add_sleep = TRUE, show_progress = TRUE)
}
\arguments{
\item{year}{integer for year of contest. Currently only the
year \code{2021} is supported.}

\item{add_sleep}{boolean to add a simple sleep between the calls to
extracting contest post details.}

\item{show_progress}{boolean to show progress bar. Default is \code{TRUE}.}
}
\value{
tidy data frame with post metadata and links
}
\description{
Obtain Shiny Contest Submissions
}
