% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_post_links.R
\name{get_post_links}
\alias{get_post_links}
\title{Extract all links contained in a discourse post}
\usage{
get_post_links(url, add_sleep = TRUE, pb = NULL)
}
\arguments{
\item{url}{character string of the URL for a post.}

\item{add_sleep}{boolean to add a simple sleep before
function completes.}

\item{pb}{progress bar object. If no progress output is
needed, keep this value as \code{NULL}.}
}
\value{
character vector of all links
}
\description{
Extract all links contained in a discourse post
}
